Plates outlines finished
Redefined the various outlines of the plates based on the new constrains.
For the top dimension, I took the largest middle-finger stagerring I could find (Elephant42, at +9.5mm) and rounded it to 10mm max amplitude in the middle finger rail, and used that as reference for the position of the top arc (giving 3mm margin above that highest slot).
And defined the position of the rest from that.
Top left M% hole might need to be reviewed.