Plates: Received M2 and M5 hardware. It's bigger than specified...
- M5 screw head is ~9.5 instead of 8.7. Was colliding in the top left 
corner with M2 screw for pro micro board. So I shifted up pro micro 
board 0.3mm.
- M2 screw head is also larger than expected (3.8 instead of 3.2). But 
it is mostly fine as I was designing with 4-4.2mm clearance. Only 
exception was for screws supporting the middle bottom section of top 
plate with rails. Here it was colliding with switch housing. So, only 
solution is to use so called reduced slotted screws (DIN 920) as they 
have 2.8mm diameter head.
- Also added JLCPCB silk reference