Handy: Removing nav cluster, adding outer column
During layout tests using Lergo keeb, I figured out the nav cluster is actually confusing (beside looking ugly).

So, here it is removed. The keymap is updated to bring back Nav layer (+ some additions to it found with the Lergo tests, like bottom brough back under up on middle finger, home/end on the home row of inner/pinky, Ctrl+left/right under left and right).

While at it, I also merged Media with Nav layer. Some keys like vol. up/down, or prev/next was added around the nav keys following similar logic. Mute and stop have been reshuffled around in the thumb cluster to keep backspace and delete free when using Nav layer.

NB: Key position middle top remains free for future use.

To compensate the "loss" of modifiers keys (what used to be the mod row, which had been compensated using the nav cluster, but now need a new place), the outer column is introduced. Since it is quite new, and wasn't able to test it on the Lergo platform, I only assigned keys that seems to be usable there. But there is quite a large room for improvements and future modifications.

On the left hand there is left Ctrl on bottom row (which I really seem to not be able to get rid off...), backspace on home row (from a very appreciated Colemak mod, though here it is in "competition" with the also much appreciated thumb cluster position, but at least it is on a different hand). And left Alt on top row (since VSCode use alt a lot, let's see how that goes).

On the right hand side, GUI is on the bottom row (still needed for locking screen for instance, for which having it on the same hand than L is nice). Home row is unassigned still. And top row got quote/doublequote as a mimic of what I currently have on the constrained ANSI layout of laptops.