Finally, can manipulate KiCad files!
That's a fat one.
We now have a way to read, parse, and deeply reify KiCad files using S-Expr format.
But also re-export (ie. serialise) back into S-Expr text, striving for minimal formatting changes.

The first implemented tokens are for .kicad_pcb and .kicad_mod files, supporting the simplest items.
And the first usage is for taking the ergogen output in the form of (old) .kicad_pcb, convert them to newer format (including trigonometric recalculations for arc items), then convert all the graphic items generated to ergogen into a footprint, that is then exported into a library.

The code is ... voluminous, and far from being trivial. But it has great potential as it allows doing deep modifications to objects defined in a KiCad file, while being purely offline (ie. outside of KiCad entirely).

Some tests are included as well.