Ergogen: Adds "preprocessing" capabilites, for those really complex points
Some points in Lergo/Handy cases needs some math and trigonometry a bit more involved than what ergogen can do.

Notably, I had to implement intersection of 2 lines, intersection of a line and a  circle, and placement of a circle such that it is tangent to another circle, while its center reamains on a line.