Handy: Proper case implementation
Since there is now "preprocessin" capabilities added to ergogen by keebs.py script, it can now calculate precisely some of the difficult points (usually corners) that are defined by intersection and/or tangential logic of placement.

Tenting holes have been added as well, with an option for the "inner" top tenting screw being either near the middle (to be parallel to the right edge) or all the way to the left (vertically aligned with the bottom left one), like it is on Lergo. Will see on first prototype how the stability goes for either solution.

Also cleaned up a bit all the outlines, added premises of components, starting with promicro, oled screen and tenting hole, some debugging facilities, notably display of  construction points and orientations of some points.