Kicad: Nails further issues with angles...
Despite the special casing done in _angle_normalizer for fp_text and pad, there was still some unnecessary modifications of, eg. 355 to -5.
This was due to the way footprint are updated: for every item in the model footprint (which usually is built with a 0 degree reference), a new item with angle applied (the one of each particular ref being updated) to fp_texts and pads  is made. Then it is *compared* to the existing items we already have on the board. And if the comparison showed they are the same, then the "old" item is used (to avoid unnecessarily changing their tstamp ID).

Issue was that this new item used for comparison to the old one was having the "normally correct" angle in the -180 to 180 range, as there does not seem to be a reliable way of knowing when kicad use the bizarre -180 to 360 range for fp_text and pads.
And so, when comparing, the algorithm though they were different from the ones expressed in that bizarre range, and updated them, using the new version. And as such creating useless diffs.

Solution is to use a special angle comparison callback that will translate any angle (ie. not doing special casing for fp_text and pads here) to the normal range. Like so, in the footprint update algorithm, the new item compares as equal to the old one, despite using different angle values (but still truly the same on unit circle).

In that process, I figured the way angles were normalized outside of one of those two ranges was actually mathematically wrong. It just happened that all instances I had in the tests for this case ended up being coincidently right despite the wrong formula...
So, now at least the simple _angle_key function is more properly tested. And changes were propagated to _angle_normalizer (though that one would also need proper testing one day, maybe once I know for sure how that bizarre range works, and if it is really -180 to 360 and not something else...).