KiCad: Fixes bug introduced by one speed optimisation
In case of loading different kicad_pcb files of different versions within the same instance of a script, some versioned token of the wrong version were weirdly used. But if parsing each file independently then it was fine...

Turns out, the method doing the resolution of versioned token was cached in such a way that if later it was used in the context of a different version, then it was using the cached result from the first time it was used...

Taking the version context out of that method, and instead passing the current version as an argument to it fixes the issue.